+++ src/components/organisms/header/index.tsx
@@ -1,0 +1,5 @@
+import Logo from '@/components/organisms/header/Logo';
+import GNB from '@/components/organisms/header/GNB';
+import UserMenu from '@/components/organisms/header/UserMenu';
+
+export { GNB, Logo, UserMenu };
